+++ clients/java/src/main/java/marquez/client/models/ColumnLineageNodeData.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2018-2022 contributors to the Marquez project
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package marquez.client.models;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NonNull;
+import marquez.client.Utils;
+
+@Getter
+@AllArgsConstructor
+@EqualsAndHashCode
+public class ColumnLineageNodeData implements NodeData {
+  @NonNull String namespace;
+  @NonNull String dataset;
+  @NonNull String field;
+  @NonNull String fieldType;
+  @NonNull String transformationDescription;
+  @NonNull String transformationType;
+  @NonNull List<DatasetFieldId> inputFields;
+
+  public static ColumnLineageNodeData fromJson(@NonNull final String json) {
+    return Utils.fromJson(json, new TypeReference<ColumnLineageNodeData>() {});
+  }
+}